Moved the event form to a common component
