+++ frontend/src/common/components/event-form/index.jsx
@@ -1,0 +1,3 @@
+import EventForm from "./event.form";
+
+export default EventForm;
